requirements.txt
pyzmq>=25.0.0
websockets>=12.0
redis>=5.0.0
numpy>=1.24.0
